log-monitoring-agent/requirements.txt
openai>=1.30.0
python-dotenv